src: Move device info functionality from unit tests
