+++ src/stdgpu/openmp/CMakeLists.txt
@@ -5,7 +5,8 @@ set(STDGPU_DEPENDENCIES_BACKEND_INIT "
 find_dependency(OpenMP 2.0 REQUIRED)
 " PARENT_SCOPE)
 
-target_sources(stdgpu PRIVATE impl/memory.cpp)
+target_sources(stdgpu PRIVATE impl/device.cpp
+                              impl/memory.cpp)
 
 target_compile_definitions(stdgpu PUBLIC THRUST_DEVICE_SYSTEM=THRUST_DEVICE_SYSTEM_OMP)
