+++ src/stdgpu/openmp/device.h
@@ -1,0 +1,35 @@
+/*
+ *  Copyright 2019 Patrick Stotko
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+#ifndef STDGPU_OPENMP_DEVICE_H
+#define STDGPU_OPENMP_DEVICE_H
+
+namespace stdgpu
+{
+
+namespace openmp
+{
+
+/**
+ * \brief Prints the technical data of the currently used device
+ */
+void
+print_device_information();
+
+} // namespace openmp
+
+} // namespace stdgpu
+
+#endif // STDGPU_OPENMP_DEVICE_H
